requirements.txt
# API Framework
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Utils
requests>=2.31.0
python-multipart>=0.0.6

# Fetch.ai
requests>=2.31.0
uagents>=0.6.0

# For testing
pytest>=7.4.2
pytest-asyncio>=0.21.1

# Watchdog
watchdog>=2.3.1 